requirements.txt
openai
tiktoken
python-dotenv
langchain